+++ demo/api/src/main.ts
@@ -1,3 +1,5 @@
+import helmet from "helmet";
+
 if (process.env.TRACING_ENABLED) {
     require("./tracing");
 }
@@ -39,6 +41,18 @@ async function bootstrap(): Promise<void> {
         }),
     );
 
+    app.use(
+        helmet({
+            contentSecurityPolicy: {
+                useDefaults: false,
+                directives: {
+                    "default-src": helmet.contentSecurityPolicy.dangerouslyDisableDefaultSrc,
+                    // locally: allow localhost in frame-ancestors to enable including files from the API in iframes in admin
+                    "frame-ancestors": `'self' ${process.env.NODE_ENV === "development" ? config.adminUrl : ""}`,
+                },
+            },
+        }),
+    );
     app.use(json({ limit: "1mb" })); // increase default limit of 100kb for saving large pages
     app.use(compression());
     app.use(cookieParser());