+++ packages/api/cms-api/src/dam/files/files.resolver.ts
@@ -1,16 +1,17 @@
 import { InjectRepository } from "@mikro-orm/nestjs";
 import { EntityRepository } from "@mikro-orm/postgresql";
 import { Inject, NotFoundException, Type } from "@nestjs/common";
-import { Args, ID, Mutation, ObjectType, Parent, Query, ResolveField, Resolver } from "@nestjs/graphql";
+import { Args, Context, ID, Mutation, ObjectType, Parent, Query, ResolveField, Resolver } from "@nestjs/graphql";
+import { IncomingMessage } from "http";
 import { basename, extname } from "path";
 
-import { CurrentUserInterface } from "../../auth/current-user/current-user";
 import { GetCurrentUser } from "../../auth/decorators/get-current-user.decorator";
 import { SkipBuild } from "../../builds/skip-build.decorator";
 import { CometValidationException } from "../../common/errors/validation.exception";
 import { PaginatedResponseFactory } from "../../common/pagination/paginated-response.factory";
 import { AffectedEntity } from "../../user-permissions/decorators/affected-entity.decorator";
 import { RequiredPermission } from "../../user-permissions/decorators/required-permission.decorator";
+import { CurrentUser } from "../../user-permissions/dto/current-user";
 import { ACCESS_CONTROL_SERVICE } from "../../user-permissions/user-permissions.constants";
 import { AccessControlServiceInterface } from "../../user-permissions/user-permissions.types";
 import { DAM_FILE_VALIDATION_SERVICE } from "../dam.constants";
@@ -118,7 +119,7 @@ export function createFilesResolver({ File, Scope: PassedScope }: { File: Type<F
         @SkipBuild()
         async moveDamFiles(
             @Args({ type: () => MoveDamFilesArgs }) { fileIds, targetFolderId }: MoveDamFilesArgs,
-            @GetCurrentUser() user: CurrentUserInterface,
+            @GetCurrentUser() user: CurrentUser,
         ): Promise<FileInterface[]> {
             let targetFolder = null;
             if (targetFolderId !== null) {
@@ -147,7 +148,7 @@ export function createFilesResolver({ File, Scope: PassedScope }: { File: Type<F
         @Mutation(() => CopyFilesResponse)
         @SkipBuild()
         async copyFilesToScope(
-            @GetCurrentUser() user: CurrentUserInterface,
+            @GetCurrentUser() user: CurrentUser,
             @Args("fileIds", { type: () => [ID] }) fileIds: string[],
             @Args("inboxFolderId", {
                 type: () => ID,
@@ -256,8 +257,8 @@ export function createFilesResolver({ File, Scope: PassedScope }: { File: Type<F
         }
 
         @ResolveField(() => String)
-        async fileUrl(@Parent() file: FileInterface): Promise<string> {
-            return this.filesService.createFileUrl(file);
+        async fileUrl(@Parent() file: FileInterface, @Context("req") req: IncomingMessage): Promise<string> {
+            return this.filesService.createFileUrl(file, Boolean(req.headers["x-preview-dam-urls"]));
         }
 
         @ResolveField(() => [File])